[WIP] rewrite analyzer to allow line rewriting
This makes it possible to have completely different
lines displayed than the ones received

(use case: receive command output in JSON)
