+++ src/analysis/analyzer.rs
@@ -1,8 +1,9 @@
 use {
+    anyhow::Result,
     super::{
         eslint,
         nextest,
-        python,
+        //python,
         standard,
     },
     crate::*,
@@ -17,7 +18,7 @@ use {
 /// Implementation routing will probably change at some point
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub enum Analyzer {
+pub enum AnalyzerRef {
     #[default]
     Standard,
     Nextest,
@@ -25,30 +26,64 @@ pub enum Analyzer {
     PythonUnittest,
 }
 
-impl Analyzer {
-    pub fn analyze_line(
+impl AnalyzerRef {
+
+    pub fn create_analyzer(
         self,
-        line: &CommandOutputLine,
-    ) -> LineAnalysis {
+    ) -> Box<dyn Analyzer> {
         match self {
-            Self::Eslint => eslint::analyze_line(line),
-            Self::Standard => standard::analyze_line(line),
-            Self::Nextest => nextest::analyze_line(line),
-            Self::PythonUnittest => python::unittest::analyze_line(line),
-        }
-    }
-    pub fn build_report(
-        &self,
-        cmd_lines: &[CommandOutputLine],
-        mission: &Mission,
-    ) -> anyhow::Result<Report> {
-        match self {
-            Self::Eslint => eslint::build_report(cmd_lines, *self, mission),
-            Self::PythonUnittest => python::unittest::build_report(cmd_lines, *self, mission),
-            Self::Standard | Self::Nextest => {
-                // nextest analyzis simply uses the standard report building
-                standard::build_report(cmd_lines, *self, mission)
+            Self::Standard => Box::new(standard::StandardAnalyzer::default()),
+            Self::Nextest => Box::new(nextest::NextestAnalyzer::default()),
+            Self::Eslint => Box::new(eslint::EslintAnalyzer::default()),
+            _ => {
+                todo!()
             }
         }
     }
+
+}
+
+pub trait Analyzer {
+
+    fn start(
+        &mut self,
+        mission: &Mission,
+    );
+
+    fn receive_line(
+        &mut self,
+        line: CommandOutputLine,
+        command_output: &mut CommandOutput,
+    ); // return line type ? analysis ?
+
+    fn build_report(
+        &mut self,
+    ) -> Result<Report>;
+
+
+    //pub fn analyze_line(
+    //    self,
+    //    line: &CommandOutputLine,
+    //) -> LineAnalysis {
+    //    match self {
+    //        Self::Eslint => eslint::analyze_line(line),
+    //        Self::Standard => standard::analyze_line(line),
+    //        Self::Nextest => nextest::analyze_line(line),
+    //        Self::PythonUnittest => python::unittest::analyze_line(line),
+    //    }
+    //}
+    //pub fn build_report(
+    //    &self,
+    //    cmd_lines: &[CommandOutputLine],
+    //    mission: &Mission,
+    //) -> anyhow::Result<Report> {
+    //    match self {
+    //        Self::Eslint => eslint::build_report(cmd_lines, *self, mission),
+    //        Self::PythonUnittest => python::unittest::build_report(cmd_lines, *self, mission),
+    //        Self::Standard | Self::Nextest => {
+    //            // nextest analyzis simply uses the standard report building
+    //            standard::build_report(cmd_lines, *self, mission)
+    //        }
+    //    }
+    //}
 }
