+++ src/analysis/eslint/mod.rs
@@ -1,42 +1,37 @@
 //! An analyzer for eslint ( https://eslint.org/ )
+mod eslint_line_analyzer;
+
 use {
     super::*,
     crate::*,
-    lazy_regex::*,
+    anyhow::Result,
+    eslint_line_analyzer::*,
 };
 
-const CSI_LOCATION_PATH: &str = "\u{1b}[4m";
-const CSI_LINE_COL: &str = "\u{1b}[2m";
-const CSI_ERROR: &str = "\u{1b}[31m";
-const CSI_WARNING: &str = "\u{1b}[33m";
-const CSI_SUM: &str = "\u{1b}[31m\u{1b}[1m";
+#[derive(Debug, Default)]
+pub struct EslintAnalyzer {
+    lines: Vec<CommandOutputLine>,
+}
 
-pub fn analyze_line(cmd_line: &CommandOutputLine) -> LineAnalysis {
-    if is_line_col_error_line(&cmd_line.content) {
-        LineAnalysis {
-            line_type: LineType::Title(Kind::Error),
-            key: None,
-        }
-    } else if is_line_col_warning_line(&cmd_line.content) {
-        LineAnalysis {
-            line_type: LineType::Title(Kind::Warning),
-            key: None,
-        }
-    } else if get_location_path(&cmd_line.content).is_some() {
-        LineAnalysis {
-            line_type: LineType::Location,
-            key: None,
-        }
-    } else if is_sum(&cmd_line.content) {
-        LineAnalysis {
-            line_type: LineType::Title(Kind::Sum),
-            key: None,
-        }
-    } else {
-        LineAnalysis {
-            line_type: LineType::Normal,
-            key: None,
-        }
+impl Analyzer for EslintAnalyzer {
+
+    fn start(&mut self, _mission: &Mission) {
+        self.lines.clear();
+    }
+
+    fn receive_line(
+        &mut self,
+        line: CommandOutputLine,
+        command_output: &mut CommandOutput,
+    ) {
+        self.lines.push(line.clone());
+        command_output.push(line);
+    }
+
+    fn build_report(
+        &mut self,
+    ) -> Result<Report> {
+        build_report(&mut self.lines)
     }
 }
 
@@ -47,20 +42,11 @@ pub fn analyze_line(cmd_line: &CommandOutputLine) -> LineAnalysis {
 /// in the file.
 pub fn build_report(
     cmd_lines: &[CommandOutputLine],
-    line_analyzer: Analyzer,
-    mission: &Mission,
 ) -> anyhow::Result<Report> {
-    let ignore_patterns = mission.ignored_lines_patterns();
+    let line_analyzer = EslintLineAnalyzer{};
     let mut items = ItemAccumulator::default();
     let mut last_location_path = None;
     for cmd_line in cmd_lines {
-        if let Some(patterns) = ignore_patterns {
-            let raw_line = cmd_line.content.to_raw();
-            if patterns.iter().any(|p| p.raw_line_is_match(&raw_line)) {
-                debug!("ignoring line: {}", &raw_line);
-                continue;
-            }
-        }
         let line_analysis = line_analyzer.analyze_line(cmd_line);
         let line_type = line_analysis.line_type;
         match line_type {
@@ -115,90 +101,3 @@ pub fn build_report(
     Ok(report)
 }
 
-/// Return true when the line is like
-///    "67:52  error  Unnecessary escape character: \/  no-useless-escape"
-fn is_line_col_error_line(content: &TLine) -> bool {
-    let mut strings = content.strings.iter();
-    let (Some(first), Some(second), Some(third), Some(fourth)) = (
-        strings.next(),
-        strings.next(),
-        strings.next(),
-        strings.next(),
-    ) else {
-        return false;
-    };
-    first.is_blank()
-        && second.csi == CSI_LINE_COL
-        && regex_is_match!(r"^\d+:\d+$", &second.raw)
-        && third.is_blank()
-        && fourth.csi == CSI_ERROR
-        && fourth.raw == "error"
-}
-
-/// Return true when the line is like
-///    "67:52  warning bla bla"
-fn is_line_col_warning_line(content: &TLine) -> bool {
-    let mut strings = content.strings.iter();
-    let (Some(first), Some(second), Some(third), Some(fourth)) = (
-        strings.next(),
-        strings.next(),
-        strings.next(),
-        strings.next(),
-    ) else {
-        return false;
-    };
-    first.is_blank()
-        && second.csi == CSI_LINE_COL
-        && regex_is_match!(r"^\d+:\d+$", &second.raw)
-        && third.is_blank()
-        && fourth.csi == CSI_WARNING
-        && fourth.raw == "warning"
-}
-
-fn is_sum(content: &TLine) -> bool {
-    let mut strings = content.strings.iter();
-    let Some(first) = strings.next() else {
-        return false;
-    };
-    if !(first.csi == CSI_SUM
-        && regex_is_match!(r"^✖ \d+ problems \(\d+ errors, \d+ warnings\)$", &first.raw))
-    {
-        return false;
-    }
-    for string in strings {
-        if !string.is_blank() {
-            return false;
-        }
-    }
-    true
-}
-fn get_location_path(content: &TLine) -> Option<String> {
-    let mut strings = content.strings.iter();
-    let first = strings.next()?;
-    if first.csi != CSI_LOCATION_PATH {
-        return None;
-    }
-    // trying to recognize a path, I might make some wrong assumptions here,
-    // especially for windows...
-    if !regex_is_match!(r"^\s*/\S+\.\w+s\s*$", &first.raw) {
-        return None;
-    }
-    Some(first.raw.to_string())
-}
-
-fn cleaned_tline(content: &TLine) -> TLine {
-    let mut tline = TLine::default();
-    let mut last_is_blank = true;
-    for ts in &content.strings {
-        if ts.csi == CSI_LINE_COL && regex_is_match!(r"^\d+:\d+$", &ts.raw) {
-            continue; // useless line:col at start of title
-        }
-        let raw = regex_replace_all!(r"\s+", &ts.raw, " ").to_string();
-        let is_blank = raw.trim().is_empty();
-        if !(is_blank && last_is_blank) {
-            tline.strings.push(TString::new(&ts.csi, raw));
-        }
-        last_is_blank = is_blank;
-    }
-    tline
-}
