+++ src/analysis/mod.rs
@@ -5,9 +5,10 @@ mod line_analysis;
 mod line_pattern;
 mod line_type;
 mod nextest;
-mod python;
+//mod python;
 mod standard;
 mod stats;
+mod line_analyzer;
 
 pub use {
     analyzer::*,
@@ -16,4 +17,5 @@ pub use {
     line_pattern::*,
     line_type::*,
     stats::*,
+    line_analyzer::*,
 };
