+++ src/result/command_result.rs
@@ -23,12 +23,9 @@ impl CommandResult {
     pub fn build(
         output: CommandOutput,
         exit_status: Option<ExitStatus>,
-        analyzer: Analyzer,
-        mission: &Mission,
+        mut report: Report,
     ) -> Result<Self> {
-        let lines = &output.lines;
         let error_code = exit_status.and_then(|s| s.code()).filter(|&c| c != 0);
-        let mut report = analyzer.build_report(lines, mission)?;
         debug!("report stats: {:?}", &report.stats);
         if let Some(error_code) = error_code {
             let stats = &report.stats;
