+++ src/result/mod.rs
@@ -5,8 +5,10 @@ mod line;
 mod report;
 mod wrapped_command_output;
 mod wrapped_report;
+mod report_maker;
 
 pub use {
+    report_maker::*,
     command_output::*,
     command_result::*,
     failure::*,
